Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, FlatList, Text } from 'react-native';
+import { View, FlatList, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -20,15 +20,30 @@ const mapDispatchToProps = dispatch => ({
 
 class Favorites extends Component {
     render () {
-        const { navigate } = this.props.navigation;
-
-        
+        const { navigate } = this.props.navigation;        
         
         const renderMenuItem = ({ item, index }) => {
             const rightButton = [{
                 text: 'Delete',
                 type: 'delete',
-                onPress: () => this.props.deleteFavorite(item.id)
+                onPress: () => {
+                    Alert.alert(
+                        'Delete Favorite?',
+                        'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                        [
+                            {
+                                text: 'Cancel',
+                                onPress: () => console.log(item.name + ' not deleted!'),
+                                style: 'cancel'
+                            },
+                            {
+                                text: 'OK',
+                                onPress: () => this.props.deleteFavorite(item.id)
+                            }
+                        ],
+                        { cancelable: false }
+                    );
+                }
             }];
 
             return (
